src/components/private-route/private-route.tsx
import { Navigate } from 'react-router-dom';
import { useAppSelector } from '../../hooks';
import { getAuthorizationStatus } from '../../store/user-process/types-user-process';
import { AppRoute, AuthorizationStatus } from '../../utils/const';

type PrivateRouteProps = {
  children: JSX.Element;
}

function PrivateRoute ({children}: PrivateRouteProps): JSX.Element {
  const authorizationStatus = useAppSelector(getAuthorizationStatus);
  return (
    authorizationStatus === AuthorizationStatus.Auth ? children : <Navigate to={AppRoute.Login} />
  );
}

export default PrivateRoute;
